refactor(map): 地図画面の位置情報ロジックをカスタムフックに分離
map-top.tsxの責務を分離し、保守性と再利用性を向上

- useLocation.ts: 位置情報管理ロジック（権限・取得・追従）
- useMapCamera.ts: 地図カメラ制御ロジック（追従・操作検知）
- LocationBanner/RecenterButton: UIコンポーネント分離
- 型定義・定数・サンプルデータを別ファイルに移動

map-top.tsx: 530行 → 204行（326行削減、61.5%減）

Closes #48
