+++ docs/issue/#48/overview.md
@@ -60,3 +60,50 @@
     // ...
   }
   推定削減: 約150行
+
+  ---
+  ## 実装完了報告
+
+  ### 実装日時
+  2025年10月21日
+
+  ### 実装内容
+  フェーズ0〜3を完了し、map-top.tsxの責務分離とモジュール化を実現
+
+  #### 成果物
+  1. **新規作成ファイル（7ファイル）**
+     - `app/types/location.ts` - LocationStatus型定義
+     - `app/constants/colors.ts` - TAB_COLOR等の定数
+     - `app/data/sampleRoutes.ts` - サンプルルートデータ
+     - `app/components/map/LocationBanner.tsx` - 位置情報バナーUI（98行）
+     - `app/components/map/RecenterButton.tsx` - 現在地ボタンUI（41行）
+     - `app/hooks/useMapCamera.ts` - 地図カメラ制御ロジック（117行）
+     - `app/hooks/useLocation.ts` - 位置情報管理ロジック（175行）
+
+  2. **修正ファイル**
+     - `app/screens/map-top.tsx` - 530行 → 204行（**326行削減、61.5%減**）
+     - `CLAUDE.md` - モジュール構成の説明を更新
+
+  #### 数値成果
+  | 項目 | Before | After | 改善率 |
+  |------|--------|-------|--------|
+  | map-top.tsx行数 | 530行 | 204行 | **-61.5%** |
+  | 責務の数 | 4つ（密結合） | 1つ（地図表示） | **-75%** |
+  | モジュール数 | 1ファイル | 8ファイル | +700% |
+  | 再利用可能性 | ❌ | ✅ | 質的向上 |
+
+  #### アーキテクチャ改善
+  - **UI層**: LocationBanner, RecenterButton（プレゼンテーショナル）
+  - **ロジック層**: useLocation, useMapCamera（カスタムフック）
+  - **データ層**: types, constants, data（型定義・定数・サンプル）
+
+  ### 完了条件チェック
+  - [x] 分離したモジュールが動作する（各モジュールがmap-top.tsxで正常に動作）
+  - [x] map-top.tsxの行数が削減（530→204行、326行削減）
+  - [x] Issue #25で再利用可能（useLocation.tsをそのまま利用可能）
+  - [ ] 手動チェックリストで確認済み（実機テスト待ち）
+
+  ### 次のステップ
+  1. 実機での動作確認（位置情報権限フロー、追従、再センタリング）
+  2. Issue #25「ルート提案画面」でuseLocation.tsを再利用
+  3. 必要に応じてユニットテスト追加
